pyproject.toml
[project]
name = "your-stretch"
version = "0.1.0"
description = ""
authors = [
    {name = "ToHellForever",email = "dim.anosoff2018@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
